Implement GitHub Pages deployment via API
The old strategy is moved to `git`, but it is not fully tested.
