+++ lib/dpl/provider.rb
@@ -166,7 +166,7 @@ module Dpl
       :validate_runtimes, :user_agent
 
     def_delegators :ctx, :apt_get, :gem_require, :npm_install, :pip_install,
-      :build_dir, :build_number, :repo_slug, :encoding, :git_commit_msg,
+      :build_dir, :build_number, :repo_slug, :encoding, :git_branch, :git_commit_msg,
       :git_dirty?, :git_log, :git_ls_files, :git_ls_remote?, :git_remote_urls,
       :git_rev_parse, :git_sha, :git_tag, :machine_name, :node_version,
       :npm_version, :sleep, :ssh_keygen, :success?, :mv, :tmp_dir, :which,
